windows: Better error message on failed commit
When importing a layer, prefer the error message from the
layer reader to that of the layer writer, since this is
more likely to be a useful message. This is achieved by
always returning a zero exit code from re-exec, even on error.
This makes cmd.Wait() return the result of the io.Copy used
to copy in the layer data rather than the ExitError from
the process returning a non-zero exit code.

Signed-off-by: John Starks <jostarks@microsoft.com>